// src/type.d.ts
export interface ITodo {
    _id: string
    name: string
    description: string
    status: boolean
    createdAt?: string
    updatedAt?: string
  }
  
 export interface TodoProps {
    todo: ITodo
  }
  
 export type ApiDataType = {
    message: string
    status: string
    todos: ITodo[]
    todo?: ITodo
  }